fix(platform): wire nvm_store_init() into real Zephyr init paths
Addresses #200 (kept open — see issue comment: fixed and verified
against every CI-buildable Zephyr target here, real-hardware validation
still pending before final close).

nvm_store_init() was defined but never called anywhere in a real Zephyr
init path -- not zephyr_port_init(), not any example's main(), nowhere
outside unit tests. nvm_store_is_ready() therefore returned false
forever on every Zephyr build: DTC mirror persistence (already
documented as surviving power cycles) and security counter/lockout
persistence (once wired, #190) both silently never activated. Both
degrade gracefully rather than crash, which is exactly why this went
unnoticed. FreeRTOS was unaffected -- eds_platform_init() already calls
nvm_store_init().

Wired the missing call into main(), before uds_generated_init() (whose
Step 3.5 calls dtc_mirror_init(), which needs nvm_store ready):

- basic_ecu: the one example CI actually cross-compiles for 4 different
  board targets (native_sim + nucleo_h743zi + frdm_mcxn947 + mr_canhubk3,
  each with its own diag_nvs DTS partition and physical erase-page size:
  128 KB / 128 KB / 8 KB respectively). Resolves
  FIXED_PARTITION_DEVICE/OFFSET/SIZE(eds_nvs_slot) -- the DTS *node
  label*, not the partition's string "label" property ("diag_nvs"); the
  string form compiles-referenced but fails to link with an undeclared
  __device_dts_ord_DT_N_NODELABEL_diag_nvs symbol (caught in this PR's
  own CI, all three real boards, on an earlier push) -- and queries the
  real sector size at runtime via flash_get_page_info_by_offs() rather
  than hardcoding one, since a single hardcoded value cannot be correct
  across boards with such different physical sector sizes.

  This surfaced a real latent bug: nvm_store_cfg_t.sector_size was
  uint16_t, which silently truncates 128 KB (0x20000) to 0 -- would have
  broken nvs_mount() on 2 of the 3 real boards. Widened to uint32_t
  (platform/nvm_store.h); s_nvs_fs.sector_size was already uint32_t
  internally, so this is a pure widening, no other call site affected.

  mr_canhubk3 (NXP S32K344) is a further, deliberate exception even with
  the node-label fix: it still fails to *link*
  ('__device_dts_ord_199' undeclared in FIXED_PARTITION_DEVICE's
  expansion, also caught in this PR's own CI) -- &flash0 has a
  devicetree node on this SoC but no instantiated struct device behind
  it in Zephyr v3.7.0, unlike the other two real boards where the
  identical pattern links cleanly. Root-causing an NXP S32K3 flash
  driver gap needs real hardware or S32K3 HAL knowledge this repo
  doesn't have, so nvm_store_init(NULL) is used on this one board only
  (compiles, and honestly reports UDS_STATUS_ERR_NULL_PTR every boot
  rather than silently succeeding) -- same category of gap as ardep_ecu
  below, not guessed at a third time.

- safeboot_ecu: its board's diag_nvs partition and exact sector geometry
  are already fully documented in
  boards/nucleo_h743zi/nucleo_h743zi.overlay (added for a different
  purpose earlier) -- wired directly from that (same eds_nvs_slot node
  label fix as basic_ecu). Not currently compiled by CI (example-safeboot
  only validates generated files), so this one is the most in need of the
  real-hardware validation this issue stays open for.

- basic_ecu_doip, bms_ecu, motor_controller_ecu,
  robot_joint_controller_ecu, sensor_ecu: native_sim-only targets, where
  platform/zephyr/nvm_store_mock.c (RAM-backed) is always linked
  regardless of what's passed -- nvm_store_init(NULL) is correct and
  sufficient there (nvm_store.h: "may be NULL on host").

ardep_ecu is a deliberate, documented exception: CONFIG_NVS=y is set but
no diag_nvs partition exists in its overlay, and the ARDEP board DTS
itself lives in an external, non-vendored Zephyr module. Guessing a
flash layout for hardware this repo has no visibility into is exactly
the risk this issue originally flagged against -- left unfixed rather
than guessed. Not compiled by CI either way.

Verified: bash build_tests.sh (45 passed, 912 cases, 0 failed -- these
files aren't part of the host test build, so this is a no-regression
check, not a positive verification of the new code). This repo's own
Zephyr CI cross-compile jobs are the real verification for this fix --
after two iterations catching the node-label bug and the mr_canhubk3
link gap above, native_sim, nucleo_h743zi, and frdm_mcxn947 all build
clean; mr_canhubk3 is a documented, deliberate remaining gap (see
CHANGELOG.md).

A separate, unrelated CI flake was found and filed while verifying this
PR: the "Static Analysis + MISRA C:2012" job non-deterministically
reports 6 open findings in three files this PR does not touch, given
byte-identical source and the same cppcheck version as a passing run
minutes earlier on main. See #206 -- not fixed here, out of scope.

+++ platform/nvm_store.h
@@ -117,8 +117,13 @@ typedef struct nvm_store_cfg {
     /** Flash storage offset (NVS sector start, from DTS). */
     uint32_t    flash_offset;
 
-    /** NVS sector size in bytes (typically 4096 for most flash). */
-    uint16_t    sector_size;
+    /** NVS sector size in bytes (typically 4096 for most flash).
+     *  [EDS#200] uint32_t, not uint16_t: some real targets' physical erase
+     *  pages exceed 65535 bytes (e.g. STM32H743ZI and NXP MCX N947 both
+     *  use 128 KB = 0x20000 sectors — see boards/nucleo_h743zi/*.overlay
+     *  and boards/frdm_mcxn947/*.overlay). A uint16_t here silently
+     *  truncated 0x20000 to 0, which nvs_mount() would reject. */
+    uint32_t    sector_size;
 
     /** Number of NVS sectors to use for diagnostics NVM (minimum 2). */
     uint8_t     sector_count;
